refactor: rename all source files to use consistent package name `slack`
- Replaced old `provider` package references with unified `slack` package across all source files
- Ensures consistency and avoids import conflicts during build
- No changes in provider behavior or API surface
