+++ slack/data_channels_list.go
@@ -1,4 +1,4 @@
-package provider
+package slack
 
 import (
 	"context"
