+++ slack/data_slack_user.go
@@ -1,4 +1,4 @@
-package provider
+package slack
 
 import (
 	"context"
