+++ slack/data_slack_users_group_test.go
@@ -1,4 +1,4 @@
-package provider
+package slack
 
 import "testing"
