+++ slack/resource_channel.go
@@ -1,4 +1,4 @@
-package provider
+package slack
 
 import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
