+++ slack/sync_channel_members.go
@@ -1,4 +1,4 @@
-package provider
+package slack
 
 import (
 	"fmt"
